Tons of bug fixing but UI now works properly

+++ Models/Product.cs
@@ -7,27 +7,27 @@ namespace Models
     {
         public int ID {get; set;}
         public string Name { get; set; }
-        public double Price {get; set;}
+        public decimal Price {get; set;}
 
         public Product()
         {
 
         }
         
-        public Product(string name, double price)
+        public Product(string name, decimal price)
         {
             Name = name;
             Price = price;
         }
 
-        public Product(int id, string name, double price)
+        public Product(int id, string name, decimal price)
         {
             ID = id;
             Name = name;
             Price = price;
         }
 
-        public void UpdatePrice(double price)
+        public void UpdatePrice(decimal price)
         {
             if(price < 0)
             {
